v 1.0 (Movement glitch)
I finally did it. NO MORE MOVEMENT GLITCH WHERE THEY STOP. I did this with a subroutine in the Consumer script where it just sets running away to true temporarily, and literally like that solved. The issue originated as once it ate the animal, it set object interested in to null while it was still interested in it. BUT YES. It's now at a level for data harvesting.